+++ src/protocol/codec/ext.c
@@ -79,7 +79,7 @@ z_result_t _z_msg_ext_decode_zbuf_na(_z_msg_ext_zbuf_t *ext, _z_zbuf_t *zbf) {
 z_result_t _z_msg_ext_encode(_z_wbuf_t *wbf, const _z_msg_ext_t *ext, bool has_next) {
     z_result_t ret = _Z_RES_OK;
 
-    _Z_RETURN_IF_ERR(_z_wbuf_write(wbf, _Z_EXT_FULL_ID(ext->_header) | (has_next << 7)))
+    _Z_RETURN_IF_ERR(_z_wbuf_write(wbf, (uint8_t)(_Z_EXT_FULL_ID(ext->_header) | (has_next << 7))))
 
     uint8_t enc = _Z_EXT_ENC(ext->_header);
     switch (enc) {
